fix: fix the issue of incorrect menu item value
