+++ packages/design/src/components/menu/index.module.less
@@ -14,7 +14,7 @@
 
     &-vertical {
         display: grid;
-        gap: var(--margin-xxs);
+        gap: var(--margin-xs);
     }
 
     &-hidden,
@@ -25,7 +25,7 @@
     &-item {
         &-group {
             &:not(:last-child) {
-                padding-bottom: var(--padding-xs);
+                padding-bottom: var(--padding-sm);
                 border-bottom: 1px solid rgb(var(--border-color));
             }
         }
