+++ packages/ui-plugin-sheets/src/components/menu-item-input/index.module.less
@@ -2,4 +2,8 @@
     display: flex;
     gap: var(--margin-sm);
     align-items: center;
+
+    &-container {
+        width: 72px;
+    }
 }
