+++ packages/ui-plugin-sheets/src/controllers/menu/sheet.menu.ts
@@ -1,0 +1,117 @@
+import {
+    CopySheetCommand,
+    RemoveSheetCommand,
+    SetTabColorCommand,
+    SetWorksheetHideCommand,
+    SetWorksheetShowCommand,
+} from '@univerjs/base-sheets';
+import { IMenuButtonItem, IMenuSelectorItem, MenuItemType } from '@univerjs/base-ui';
+import { ICommandService, IUniverInstanceService } from '@univerjs/core';
+import { IAccessor } from '@wendellhu/redi';
+import { Observable } from 'rxjs';
+
+import { RenameSheetOperation } from '../../commands/commands/rename.command';
+import { ShowMenuListCommand } from '../../commands/commands/unhide.command';
+import { COLOR_PICKER_COMPONENT } from '../../components/color-picker';
+import { SheetMenuPosition } from './menu';
+
+export function DeleteSheetMenuItemFactory(): IMenuButtonItem {
+    return {
+        id: RemoveSheetCommand.id,
+        type: MenuItemType.BUTTON,
+        positions: [SheetMenuPosition.SHEET_BAR],
+        title: 'sheetConfig.delete',
+    };
+}
+
+export function CopySheetMenuItemFactory(): IMenuButtonItem {
+    return {
+        id: CopySheetCommand.id,
+        type: MenuItemType.BUTTON,
+        positions: [SheetMenuPosition.SHEET_BAR],
+        title: 'sheetConfig.copy',
+    };
+}
+
+export function RenameSheetMenuItemFactory(): IMenuButtonItem {
+    return {
+        id: RenameSheetOperation.id,
+        type: MenuItemType.BUTTON,
+        positions: [SheetMenuPosition.SHEET_BAR],
+        title: 'sheetConfig.rename',
+    };
+}
+
+export function ChangeColorSheetMenuItemFactory(): IMenuSelectorItem<string> {
+    return {
+        id: SetTabColorCommand.id,
+        title: 'sheetConfig.changeColor',
+        positions: [SheetMenuPosition.SHEET_BAR],
+        type: MenuItemType.SELECTOR,
+        selections: [
+            {
+                id: COLOR_PICKER_COMPONENT,
+            },
+        ],
+    };
+}
+
+export function HideSheetMenuItemFactory(): IMenuButtonItem {
+    return {
+        id: SetWorksheetHideCommand.id,
+        type: MenuItemType.BUTTON,
+        positions: [SheetMenuPosition.SHEET_BAR],
+        title: 'sheetConfig.hide',
+    };
+}
+
+export function UnHideSheetMenuItemFactory(accessor: IAccessor): IMenuSelectorItem<any> {
+    const univerInstanceService = accessor.get(IUniverInstanceService);
+    const commandService = accessor.get(ICommandService);
+    const workbook = univerInstanceService.getCurrentUniverSheetInstance();
+    const hiddenList = workbook.getHiddenWorksheets().map((s) => ({
+        label: workbook.getSheetBySheetId(s)?.getName() || '',
+        value: s,
+    }));
+
+    return {
+        id: SetWorksheetShowCommand.id,
+        type: MenuItemType.SELECTOR,
+        positions: [SheetMenuPosition.SHEET_BAR],
+        title: 'sheetConfig.unhide',
+        disabled$: new Observable((subscriber) => {
+            const disposable = commandService.onCommandExecuted((c) => {
+                if (c.id !== SetWorksheetHideCommand.id && c.id !== SetWorksheetShowCommand.id) {
+                    return;
+                }
+                const newList = workbook.getHiddenWorksheets();
+                subscriber.next(newList.length === 0);
+            });
+            subscriber.next(hiddenList.length === 0);
+            return disposable.dispose;
+        }),
+        selections: new Observable((subscriber) => {
+            const disposable = commandService.onCommandExecuted((c) => {
+                if (c.id !== SetWorksheetHideCommand.id && c.id !== SetWorksheetShowCommand.id) {
+                    return;
+                }
+                const newList = workbook.getHiddenWorksheets().map((s) => ({
+                    label: workbook.getSheetBySheetId(s)?.getName() || '',
+                    value: s,
+                }));
+                subscriber.next(newList);
+            });
+            subscriber.next(hiddenList);
+            return disposable.dispose;
+        }),
+    };
+}
+
+export function ShowMenuItemFactory(): IMenuButtonItem {
+    return {
+        id: ShowMenuListCommand.id,
+        type: MenuItemType.BUTTON,
+        positions: [SheetMenuPosition.SHEET_BAR],
+        title: 'sheetConfig.unhide',
+    };
+}
