--- packages/ui-plugin-sheets/src/views/EditorContainer/editorContainer-zh.md
@@ -1,11 +1,0 @@
----
-category: Components
-type: 通用
-title: EditorContainer
-subtitle: 文本编辑器
-cover: ''
----
-
-## 介绍
-
-编辑器浮层，可录入公式，富文本
